task#289: move test-classes to test-root & use proper wait for C8
