+++ kadai-adapter-camunda-8-system-connector/src/test/java/io/kadai/adapter/systemconnector/camunda/api/impl/Camunda8TaskCompleterTest.java
@@ -5,7 +5,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import io.camunda.client.CamundaClient;
 import io.camunda.client.api.response.ProcessInstanceEvent;
 import io.camunda.process.test.api.CamundaAssert;
-import io.kadai.adapter.systemconnector.camunda.tasklistener.KadaiAdapterCamunda8SpringBootTest;
+
+import io.kadai.adapter.systemconnector.camunda.Camunda8TestUtil;
+import io.kadai.adapter.systemconnector.camunda.KadaiAdapterCamunda8SpringBootTest;
 import io.kadai.adapter.test.KadaiAdapterTestUtil;
 import io.kadai.common.api.KadaiEngine;
 import io.kadai.common.test.security.WithAccessId;
@@ -25,7 +27,8 @@ class Camunda8TaskCompleterTest {
   @Autowired private CamundaClient client;
   @Autowired private KadaiAdapterTestUtil kadaiAdapterTestUtil;
   @Autowired private KadaiEngine kadaiEngine;
-  @Autowired Camunda8TestUtil camunda8TestUtil;
+  @Autowired
+  Camunda8TestUtil camunda8TestUtil;
 
   @Test
   @WithAccessId(user = "admin")
